feat: minor refactor grpc insert
